+++ src/lib/components/AddressInput.tsx
@@ -4,11 +4,11 @@ import type {
   Control,
   FieldPath,
   FieldValues,
-  UseControllerProps,
+  RegisterOptions,
 } from "react-hook-form";
 import { useWatch } from "react-hook-form";
 
-import { useValidateAddress } from "lib/app-provider";
+import { useCelatoneApp, useValidateAddress } from "lib/app-provider";
 import type { FormStatus, TextInputProps } from "lib/components/forms";
 import { ControllerInput } from "lib/components/forms";
 
@@ -16,7 +16,7 @@ interface AddressInputProps<T extends FieldValues>
   extends Omit<TextInputProps, "value" | "setInputState"> {
   name: FieldPath<T>;
   control: Control<T>;
-  rules?: UseControllerProps["rules"];
+  validation?: RegisterOptions["validate"];
   maxLength?: number;
   helperAction?: ReactNode;
 }
@@ -33,13 +33,16 @@ export const AddressInput = <T extends FieldValues>({
   label,
   labelBgColor = "background.main",
   helperText,
-  placeholder = "ex.cltn1ff1asdf7988aw49efa4vw9846789",
+  placeholder,
   error,
   size = "lg",
-  rules = {},
+  validation = {},
   maxLength,
   helperAction,
 }: AddressInputProps<T>) => {
+  const {
+    appHumanAddress: { example: exampleAddr },
+  } = useCelatoneApp();
   const { validateUserAddress, validateContractAddress } = useValidateAddress();
   const validateAddress = useCallback(
     (input: string) =>
@@ -61,14 +64,17 @@ export const AddressInput = <T extends FieldValues>({
       name={name}
       control={control}
       label={label}
-      placeholder={placeholder}
+      placeholder={placeholder ?? exampleAddr}
       type="text"
       variant="floating"
       status={status}
       labelBgColor={labelBgColor}
       helperText={helperText}
       size={size}
-      rules={rules}
+      rules={{
+        required: "Address is empty",
+        validate: { validateAddress, ...validation },
+      }}
       maxLength={maxLength}
       helperAction={helperAction}
     />