+++ src/lib/components/StickySidebar.tsx
@@ -1,3 +1,4 @@
+import type { BoxProps } from "@chakra-ui/react";
 import {
   Accordion,
   AccordionButton,
@@ -8,53 +9,57 @@ import {
   Flex,
   Text,
 } from "@chakra-ui/react";
-import type { ReactElement } from "react";
+
+import { useCurrentNetwork, useInternalNavigate } from "lib/app-provider";
+import type { Network } from "lib/data";
 
 import { CustomIcon } from "./icon";
 import { SwitchToTestnet } from "./SwitchToTestnet";
 
-interface StickySidebarProps {
-  title?: string;
-  description?: ReactElement;
+export interface SidebarMetadata {
+  title: string;
+  description: React.ReactElement;
   testnetSwitch?: boolean;
-  action?: boolean;
-  marginTop?: string;
+  toStoreCode?: boolean;
 }
+
+export type SidebarDetails = Record<Network, SidebarMetadata>;
+
+interface StickySidebarProps extends BoxProps {
+  metadata: SidebarDetails;
+}
+
 export const StickySidebar = ({
-  title = " Why do I need to submit proposal?",
-  description = (
-    <span>
-      Osmosis Mainnet is permissioned chain, which means you will need to submit
-      proposal to store code.
-      <br />
-      <br /> Another way is to get your wallet address to allow list to store
-      code without opening proposal.
-      <br />
-      <br /> You still can upload your Wasm to see how your code works on
-      Testnet.
-    </span>
-  ),
-  testnetSwitch = true,
-  action = false,
-  marginTop = "36",
+  metadata,
+  ...boxProps
 }: StickySidebarProps) => {
+  const navigate = useInternalNavigate();
+  const { network } = useCurrentNetwork();
+  const { title, description, testnetSwitch, toStoreCode } = metadata[network];
+  const hasAction = testnetSwitch || toStoreCode;
   return (
-    <Box flex="4" px={8} mt={marginTop} position="relative">
-      <Flex position="fixed" width="100%">
+    <Box flex="4" px={8} position="relative" {...boxProps}>
+      <Flex position="fixed" width="full">
         <Accordion allowToggle width={96} defaultIndex={[0]}>
-          <AccordionItem>
-            <h2>
-              <AccordionButton>
-                <Box as="span" flex="1" py={2} textAlign="left">
-                  <Text variant="body2" fontWeight="600" color="text.main">
-                    {title}
-                  </Text>
-                </Box>
-                <AccordionIcon />
-              </AccordionButton>
-            </h2>
-            <AccordionPanel pb={4}>
-              <Text variant="body2" color="text.dark" pb={4}>
+          <AccordionItem borderTop="none" borderColor="pebble.700">
+            <AccordionButton py={3} px={0}>
+              <Text
+                variant="body2"
+                fontWeight={700}
+                color="text.main"
+                textAlign="start"
+              >
+                {title}
+              </Text>
+              <AccordionIcon color="pebble.600" ml="auto" />
+            </AccordionButton>
+            <AccordionPanel
+              py={3}
+              px={0}
+              borderTop="1px solid"
+              borderColor="pebble.700"
+            >
+              <Text variant="body2" color="text.dark" mb={hasAction ? 3 : 0}>
                 {description}
               </Text>
               {testnetSwitch && (
@@ -62,35 +67,43 @@ export const StickySidebar = ({
                   color="lilac.main"
                   colorHover="lilac.light"
                   bgHover="lilac.background"
-                  padding="1"
+                  p={1}
                   icon={
                     <CustomIcon
                       name="chevron-right"
                       color="lilac.main"
                       boxSize="3"
+                      m={0}
                     />
                   }
                 />
               )}
-              {action && (
-                <Text
-                  size="md"
+              {toStoreCode && (
+                <Flex
+                  align="center"
                   cursor="pointer"
                   borderRadius={8}
-                  pl="2"
+                  p={1}
+                  gap={2}
                   width="fit-content"
                   transition="all 0.25s ease-in-out"
                   color="lilac.main"
-                  py="1"
-                  _hover={{ color: "lilac.light", bgColor: "lilac.background" }}
+                  _hover={{
+                    color: "lilac.light",
+                    bgColor: "lilac.background",
+                  }}
+                  onClick={() => navigate({ pathname: "/proposal/store-code" })}
                 >
-                  Submit proposal to store code
+                  <Text variant="body3" color="inherit" fontWeight={700}>
+                    Submit Proposal To Store Code
+                  </Text>
                   <CustomIcon
                     name="chevron-right"
                     color="lilac.main"
                     boxSize={3}
+                    m={0}
                   />
-                </Text>
+                </Flex>
               )}
             </AccordionPanel>
           </AccordionItem>