+++ src/lib/components/forms/ControllerTextarea.tsx
@@ -31,7 +31,6 @@ export const ControllerTextarea = <T extends FieldValues>({
   helperText,
   placeholder = " ",
   error,
-
   height = "112px",
   rules = {},
   ...componentProps
@@ -40,13 +39,16 @@ export const ControllerTextarea = <T extends FieldValues>({
     name,
     control,
   });
-
   const { field } = useController({ name, control, rules });
-
-  const isError = !!error;
-
+  const isError = Boolean(error);
   return (
-    <FormControl size="md" isInvalid={isError} {...componentProps} {...field}>
+    <FormControl
+      size="md"
+      isInvalid={isError}
+      sx={{ "> div": { marginTop: "1 !important" } }}
+      {...componentProps}
+      {...field}
+    >
       {label && (
         <FormLabel className="textarea-label" bgColor={labelBgColor}>
           {label}