+++ src/lib/hooks/useCodeFilter.ts
@@ -2,7 +2,7 @@ import { useWallet } from "@cosmos-kit/react";
 import { useCallback } from "react";
 
 import type { CodeInfo, HumanAddr } from "lib/types";
-import { InstantiatePermission } from "lib/types";
+import { AccessConfigPermission } from "lib/types";
 
 export type PermissionFilterValue =
   | "all"
@@ -15,7 +15,7 @@ export const usePermissionFilter = (filterValue: PermissionFilterValue) => {
     ({ instantiatePermission, permissionAddresses }: CodeInfo) => {
       const isAllowed =
         permissionAddresses.includes(address as HumanAddr) ||
-        instantiatePermission === InstantiatePermission.EVERYBODY;
+        instantiatePermission === AccessConfigPermission.EVERYBODY;
 
       switch (filterValue) {
         case "with-proposal":