+++ src/pages/whitelisting.tsx
@@ -1,3 +1,0 @@
-import Whitelisting from "lib/pages/whitelisting";
-
-export default Whitelisting;